+++ src/main/java/ru/bulldog/justmap/map/waypoint/WaypointKeeper.java
@@ -12,7 +12,7 @@ import com.google.gson.JsonArray;
 import com.google.gson.JsonElement;
 import com.google.gson.JsonObject;
 
-import ru.bulldog.justmap.map.data.WorldKey;
+import ru.bulldog.justmap.map.multiworld.WorldKey;
 import ru.bulldog.justmap.util.JsonFactory;
 import ru.bulldog.justmap.util.storage.StorageUtil;
